finalisation site complet

+++ app/Http/Controllers/MessageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Message;
 use Illuminate\Support\Facades\Auth;
+// use Illuminate\Support\Facades\Gate;
 
 class MessageController extends Controller
 {
@@ -18,6 +19,8 @@ class MessageController extends Controller
     public function store(Request $request)
     {
         // validateur qui vérifie les champs
+
+        $this->authorize('create', Message::class);
 
         $request->validate([
             'contenu' => 'required|min:5|max:500',
@@ -50,6 +53,7 @@ class MessageController extends Controller
      */
     public function edit(Message $message)
     {
+        $this->authorize('update', $message);
         return \view('message.modif-message', ['message' => $message]);
     }
 
@@ -63,6 +67,12 @@ class MessageController extends Controller
      */
     public function update(Request $request, Message $message)
     {
+        // RESTRICTION D'ACCES VIA DES GATES (non utilisé car utilisation de POLICIES à la place)
+        // if (! Gate::allows('update-message', $message)) {
+        //     abort(403);
+        // }
+
+        $this->authorize('update', $message);
 
         $request->validate([
             'contenu' => 'nullable|min:5|max:500',
@@ -88,6 +98,9 @@ class MessageController extends Controller
      */
     public function destroy(Message $message)
     {
+
+        $this->authorize('delete', $message);
+
         $message->delete();
 
         return redirect()->route('home')->with('message', 'Le message a bien été supprimé');
@@ -97,6 +110,8 @@ class MessageController extends Controller
 
     function index(Request $request)
     {
+        $this->authorize('view', Message::class);
+        
         $messages = Message::where([
             [function ($query) use ($request) {
                 if (($search = $request->search)) {
